+++ Panels/Arrangers.js
@@ -64,3 +64,26 @@ export const FadeAndSlideArranger = {
 		], animationOptions);
 	}
 };
+
+/**
+ * Arranger that cross fade between panels.
+ *
+ * @type {Arranger}
+ * @private
+ */
+export const CrossFadeArranger = {
+	enter: (config) => {
+		return arrange(config, [
+			{opacity: 0, offset: 0},
+			{opacity: 0, offset: 0.5},
+			{opacity: 1, offset: 1}
+		], animationOptions);
+	},
+	leave: (config) => {
+		return arrange(config, [
+			{opacity: 1, offset: 0},
+			{opacity: 0, offset: 0.5},
+			{opacity: 0, offset: 1}
+		], animationOptions);
+	}
+};